+++ .fatherrc.ts
@@ -30,6 +30,9 @@ if (type === 'lib') {
     esm: false,
     runtimeHelpers: true,
     pkgs: [...headPkgs, ...tailPkgs],
+    extraBabelPlugins: [
+      ['babel-plugin-import', { libraryName: 'antd', libraryDirectory: 'es', style: true }, 'antd'],
+    ],
   };
 }
 
@@ -42,8 +45,8 @@ if (type === 'es') {
     runtimeHelpers: true,
     pkgs: [...headPkgs, ...tailPkgs],
     extraBabelPlugins: [
-      ['babel-plugin-import', { libraryName: 'antd', libraryDirectory: 'es', style: true }, 'antd'],
       [require('./scripts/replaceLib')],
+      ['babel-plugin-import', { libraryName: 'antd', libraryDirectory: 'es', style: true }, 'antd'],
     ],
   };
 }